test(privacy-center-config): relocate extras regression test to existing schema test home
